fix flaky test
This test was very dependent on the order tests ran in.  This caused
flakiness on Travis.  This can be demonstrated with
`@FixMethodOrder(MethodSorters.NAME_ASCENDING)` which will cause
setDefaultValues() to always fail.

The problem is that if one test mocks the values for the environment
variables (in prepareEnvMocks), then it will affect setDefaultValues
which assumes that the environment variables are unset.

The fix to this is as described in powermock/powermock#785: move the
@PrepareForTest annotation only to those tests that actually need to
mock environment variables.  That removes the flakiness.