Change deployment so that travis deploys to bintray
This allows travis to deploy to bintray.

The README explains how it should work - you tag a release and push it
to cause travis to actually deploy.  Travis checks to make sure the
git tag matches the checked-in version in gradle.properties.

We could in future look in to using something like
https://github.com/researchgate/gradle-release to automate the release
process, but this commit is the first step towards an automated CI
build.