Change deployment so that travis deploys to bintray
This removes the old way of deploying to bintray (which was manual and
handled by gradle) and instead allows travis to deploy to bintray.

The README explains how it should work - you tag a release and push it
to cause travis to actually deploy.

This commit has `dry_run: true` for testing it out.  We should remove
that and add the `on: {tags: true}` condition when we do this for real.