Remove Powermock, to enable compiling on Java11
A newer version of Powermock is required to run on Java11, but is not
available.  Newer versions of Mockito don't play well with released
versions of Powermock.

This makes changes to the tests so that they can run without
Powermock, and use a more recent version of Mockito.

Powermock was used because we need to be able to set environment
variables read via `System` to affect a static instance within the
`Configuration` class.  Further complicating matters the static
instance is generated only once, the first time it is accessed - fine
for when it is in production but not when we want to test it.  The
changes in this commit work around those issues, and this has
been tested on both Java8 and Java11.

There are two main changes in the non-test code:

* `Configuration` has a thin abstraction to read environment varibles
from `System` -> `EnvVarUtils`

* the static `EnvVarUtils` instance inside `Configuration` is replaced
with a mock in the tests.  This is possible because `Configuration`
will only initialise it once.  This enables us to override the
`EnvVarUtils` instance with a mock, clear the static `Configuration`
instance, and force `Configuration` to be re-initialised with the new
configuration set via the `EnvVarUtils` mock.

Testing `Configuration` is painful, but only because it has been
encapsulated well.

But hey, this works.