remove the MetricsBundle and update the README
I did a github code search and the only thing using MetricsBundle was
dropwizard-example, which I promptly archived.

We want people to use PrometheusBundle instead of MetricsBundle
because it does things in a nicer way.  In particular, it doesn't add
stuff that you don't necessarily need, and it binds to the admin port
instead of the application port.

We don't have a good answer (yet) for how to use PrometheusBundle on
the PaaS but (right now) I don't know if this is a pressing need for
anybody.  As a result I've removed PaaS references from the README.