G34uCPYu: add Dependabot
We want to make sure we keep our dependencies up to date. This will reduce our exposure to potential security vulnerabilities and let us fix them more quickly.

We use Dependabot for this, following the other DCS-related repositories.

We need to use the old version (1) because the new version doesn't yet support Gradle.